requirements.txt
aniso8601==10.0.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
config==0.5.1
Flask==3.1.2
Flask-HTTPAuth==4.8.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
psycopg2-binary==2.9.9
pytz==2025.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
